+++ toys-core/CHANGELOG.md
@@ -7,6 +7,7 @@
 * CHANGED: Set up a constant scope for each config file, to make constant lookup make sense.
 * CHANGED: Removed run_toys and dropped EXIT_ON_NONZERO_STATUS key in favor of using cli directly.
 * CHANGED: Renamed definition_path to source_path
+* CHANGED: LineOutput util changed to a simple Terminal util.
 * CHANGED: Organized DSL and definition classes
 * ADDED: Helper modules scoped to the tool hierarchy